Tests/LibWeb/Text/expected/hit_testing/text-with-zero-line-height.txt

anchor rect has height: true
element from point: a
click target: a
